requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-email-utils==1.0.0
django-filter==2.4.0
django-versatileimagefield==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-flex-fields==0.9.0
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
python-magic==0.4.22
python-magic-bin==0.4.14
pytz==2021.1
sqlparse==0.4.1
